requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
pillow==10.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
psutil==5.9.5
opencv-python==4.8.0.76 